Completed Companion page (1 bug)
Error - workout continuing is not permitted
 - check deleting after quitting

+++ NEA/NEA/Pages/TabbedPage/CompanionPage.xaml.cs
@@ -1,11 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using NEA.Models;
 using NEA.Models.ListViewModels;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 using NEA.Data;
 using NEA.Models.OtherModels;
+using System.Timers;
+using System.Diagnostics;
 
 namespace NEA.Pages.TabbedPage
 {
@@ -17,13 +18,50 @@ namespace NEA.Pages.TabbedPage
         Dictionary<int, int> AddedSets = new Dictionary<int, int>();
         public List<ExerciseData> Exercises;
         WorkoutRepository WorkoutRepo = new WorkoutRepository();
+        ResumeRepository ResumeRepo = new ResumeRepository();
+        private int UserID { get; set; }
+        private int Type { get; set; }
+        private string DayName { get; set; }
+        private Stopwatch Stopwatch { get; set; }
+      
 
-        public CompanionPage(List<ExerciseData> exercises)
+        public CompanionPage(List<ExerciseData> exercises, bool resume,int userID, int type, string dayname)
         {
-            WorkoutRepo.DeleteAll();
             InitializeComponent();
+            Stopwatch = new Stopwatch();
+            Stopwatch.Start();
+
             Exercises = exercises;
+
+            UserID = userID;
+            Type = type;
+            DayName = dayname;
+            NavigationPage.SetHasBackButton(this, false);
+            if (!resume)
+            {
+                WorkoutRepo.DeleteAll();
+                ResumeRepo.DeleteAll();
+            }
+            // The manually added sets for each exercise is recorded in the dictionary 
+            // In order to enable adding them to the UI
+            if (resume)
+            {
+                for (int i = 0; i < exercises.Count; i++)
+                {
+                    var check = ResumeRepo.GetAddedSets(i);
+                    if (check != 0)
+                    {
+                        AddedSets.Add(i, check);
+                    }
+                }
+            }
+            // PopulateSetLayout method adds all the entries to the UI
             PopulateSetLayout(exercises[index]);
+            // All of the previous data for sets and reps are added to the UI
+            if (resume)
+            {
+                FillEntries();
+            }
         }
         private void PopulateSetLayout(ExerciseData exercise)
         {
@@ -31,12 +69,19 @@ namespace NEA.Pages.TabbedPage
             {
                 AddedSets.Add(index, 0);
             }
+            
             Sets.Children.Clear();
             ExerciseName.Text = exercise.ExerciseName;
+
+            // For each set of the exercise, two entries are created
+            // One for recording reps and one for recording weight
+            // A change in text in the entry is recorded in the EntryData dictionary through handler methods
+            
             for (int i = 0; i < exercise.Sets + AddedSets[index]; i++)
             {
                 Entry Reps = new Entry {
                     Placeholder = exercise.Reps.ToString() + " Reps",
+                    // Keyboard is made numeric to restrict potential user error
                     Keyboard = Keyboard.Numeric
                 };
                 Reps.TextChanged += RepsEntry_TextChanged;
@@ -54,6 +99,10 @@ namespace NEA.Pages.TabbedPage
         }
         private void RepsEntry_TextChanged(object sender, TextChangedEventArgs e)
         {
+            // Exception handling is performed to prevent the program from crashing from unexpected userdata
+            // If the user inputs a non numeric value the last character is removed from the text, i.e. pressed the delete button
+            // The dictionary EntryData is updated with the changes
+            
             try
             {
                 int row = Grid.GetRow((Entry)sender);
@@ -107,11 +156,11 @@ namespace NEA.Pages.TabbedPage
                     ((Entry)sender).Text = "";
                 }
             }
-            
         }
         private void Skip_Clicked(object sender, EventArgs e)
         {
             UpdateDB();
+            /// Determines if the exercise is the last one or not
             if (index < Exercises.Count - 1)
             {
                 index++;
@@ -119,64 +168,73 @@ namespace NEA.Pages.TabbedPage
                 FillEntries();
                 Back.IsVisible = true;
             }
+            // Depending on the page the user is at the complete, and skip buttons visibility is controlled
+            // If the user is at the last exercise the complete button is shown
             if (index == Exercises.Count - 1)
             {
                 Complete.IsVisible = true;
                 Skip.IsVisible = false;
-
             }
-
+            ResumeRepo.UpdateTime(UserID, Stopwatch.Elapsed.Seconds);
         }
+
         private void FillEntries()
         {
+            // All the data for the entries for a given exercise is retrieved
             WorkoutTracker[] Logs = WorkoutRepo.GetLogs(index);
-          
-            // for each item in the log if the sets and reps arent zero then fill the entry
-            for (int i = 0; i < Exercises[index].Sets + AddedSets[index]; i++)
+            // If there is a record of data for the weight or reps the ui is updated with them
+            foreach (WorkoutTracker log in Logs)
             {
-                // if logs has an index 
-                if (Logs.Length > i)
+                if (log.Reps != 0)
                 {
-                    if (Logs[i].Reps != 0)
-                    {
-                        ((Entry)Sets.Children[i * 2]).Text = Logs[i].Reps.ToString();
-                    }
-                    if (Logs[i].Weight != 0)
-                    {
-                        ((Entry)Sets.Children[i * 2 + 1]).Text = Logs[i].Weight.ToString();
-                    }
+                    ((Entry)Sets.Children[log.Set * 2]).Text = log.Reps.ToString();
+                }
+                if (log.Weight != 0)
+                {
+                    ((Entry)Sets.Children[log.Set * 2 + 1]).Text = log.Weight.ToString();
                 }
             }
         }
         private void Back_Clicked(object sender, EventArgs e)
         {
-            if (index == Exercises.Count - 1)
-            {
-                UpdateDB();
-            }
+            
+            UpdateDB();
             index--;
             // Initially the back button is not visible to not go out of bounds for the index of the list this makes it visible
             if (index == 0)
             {
                 Back.IsVisible = false;
             }
-            // Skip is made invisible at the last exercises, if the user goes back this makes it visible again            if (index == Exercises.Count - 2)
+            // Skip is made invisible at the last exercises, if the user goes back this makes it visible again
+            if (index == Exercises.Count - 2)
             {
                 Skip.IsVisible = true;
                 Complete.IsVisible = false;
             }
             PopulateSetLayout(Exercises[index]);
             FillEntries();
+            ResumeRepo.UpdateTime(UserID, Stopwatch.Elapsed.Seconds);
         }
-        private void Quit_Clicked(object sender, EventArgs e)
+        private async void Quit_Clicked(object sender, EventArgs e)
         {
-            DisplayAlert("Quit", "Workout has been saved to resume later", "Ok");
-            Navigation.PopAsync();
-            
+            bool delete =  await DisplayAlert("Quit", "Are you sure you want to quit? all your progress will be lost", "Yes", "No");
+            // check if yes or no is clicked
+            if (delete)
+            {
+                WorkoutRepo.DeleteAll();
+                ResumeRepo.RemoveRecord(UserID);
+                ResumeRepo.DeleteAll();
+                Stopwatch.Stop();
+                ResumeRepo.ResetTime(UserID);
+                Application.Current.MainPage = new NavigationPage(new HomePage());
+            }
         }
         private void Pause_Clicked(object sender, EventArgs e)
         {
-
+            UpdateDB();
+            Stopwatch.Stop();
+            Application.Current.MainPage = new NavigationPage(new HomePage());
+    
         }
         // This method is used for adding an extra set ( with two entries for weight and reps)
         // The number of added sets are stored in the AddedSets array to be used if the back button is used
@@ -233,19 +291,65 @@ namespace NEA.Pages.TabbedPage
                 bool answer = await DisplayAlert("Incomplete", "Some sets have not been recorded, would you like to disregard them?", "Yes", "No");
                 if (answer == true)
                 {
-                   // workout is completed without changing the sets or reps  
+                    RecordProgress();
+                    ResumeRepo.DeleteAll();
+                    Stopwatch.Stop();
+                    ResumeRepo.ResetTime(UserID);
+                    ResumeRepo.RemoveRecord(UserID);
+                    Application.Current.MainPage = new NavigationPage(new HomePage());
                 }
-                else
-                {
-                    // the user is given the option to go back and fix the sets 
-                }
+            }
+            else
+            {
+                RecordProgress();
+                ResumeRepo.DeleteAll();
+                Stopwatch.Stop();
+                ResumeRepo.ResetTime(UserID);
+                ResumeRepo.RemoveRecord(UserID);
+                Application.Current.MainPage = new NavigationPage(new HomePage());
 
             }
         }
         private void UpdateDB()
         {
             WorkoutRepo.Update(EntryData, index);
+            // Check the addedsets to see if there is an entry for the index key
+            // if so updare the database
+            if (AddedSets.ContainsKey(index))
+            {
+                ResumeRepo.RecordSets(index, AddedSets[index]);
+            }
+        }
+        
+        private void RecordProgress()
+        {
+            UserDataRepository userdatarepo = new UserDataRepository();
+
+            
+            int set = 0;
+            for (int i = 0; i < Exercises.Count; i++)
+            {
+                set += Exercises[i].Sets + AddedSets[i];
+            }
+            
+            int reps = 0;
+            foreach (KeyValuePair<Tuple<int, int>, int[]> entry in EntryData)
+            {
+                reps += entry.Value[0];
+            }
+            
+            int totaltime = Stopwatch.Elapsed.Seconds;
+
+            int volume = 0;
+            foreach (KeyValuePair<Tuple<int, int>, int[]> entry in EntryData)
+            {
+                volume += entry.Value[0] * entry.Value[1];
+            }
+
+            userdatarepo.LogWorkoutData(set, reps, totaltime, volume);
 
         }
+        
+        
     }
 }
